+++ app/components/MailchimpForm.scss
@@ -37,3 +37,7 @@
     margin-bottom: 30px;
   }
 }
+
+.thanks-p {
+  margin-top: 0 !important;
+}
